移除 Compose State 从 MusicLibrary/MediaController：folderListVersion 和 metadataRefreshTrigger 改为 @Volatile var

+++ app/src/main/java/yos/music/player/ui/pages/NowPlaying.kt
@@ -248,7 +248,7 @@ fun NowPlaying(
 ) {
     val thisMusicPlaying = musicPlaying
     // 强制观察元数据刷新
-    val metaVer = metadataRefreshTrigger.intValue
+    val metaVer = metadataRefreshTrigger
     key(thisMusicPlaying.value?.title, thisMusicPlaying.value?.artists, metaVer) {
     Surface(
         modifier = Modifier.fillMaxSize(),
